backend/pyproject.toml
[project]
authors = [{ name = "James O'Claire" }]
name = "appgoblin-backend"
description = "Backend server for AppGoblin"
version = "0.0.2"

dependencies = [
    "psycopg-binary",
    "gunicorn",
    "litestar[standard]",
    "pandas",
    "numpy",
    "sqlalchemy",
    "adscrawler @ git+https://github.com/ddxv/adscrawler@main",
]

[project.optional-dependencies]
dev = ["pre-commit", "umami-analytics"]

[tool.ruff]
target-version = 'py313'
exclude = ["tests/*"]

[tool.ruff.lint]
select = ["ALL"]


ignore = [
    "W291",   # trailing space at end of line
    "G",      # Ignore pyflake logging as it does not like f-strings in log messages.
    "RET504", # Unnecessary assignment to variable before return, but I prefer for readability.
    "PD010",  # avoid pivot instead use pivot_table, great but also applies to unstack which is much more complicated
    "D211",   # Suppress ruff internal conflict
    "D213",   # Suppress ruff internal conflict
    "COM812", # ruff-format requested ignore to avoid conflict
    "D203",   # ruff-format requested ignore to avoid conflict
    "ISC001", # ruff-format requested ignore to avoid conflict
]


# Avoid trying to fix flake8-bugbear (`B`) violations.
unfixable = ["B"]

fixable = ["ALL"]


[tool.ruff.format]
docstring-code-format = true

[tool.mypy]
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true


[tool.sqlfluff.core]
templater = 'placeholder'
verbose = 0
dialect = 'postgres'
large_file_skip_byte_limit = 0
exclude_rules = 'AM04'


[tool.sqlfluff.templater.placeholder]
param_style = 'colon'
